% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{Open a connection to a SQL database, run the query, and close the connection.}
\usage{
query_db(query, config, vector = F)
}
\arguments{
\item{query}{A string containing the SQL query}

\item{config}{A list containing the parameters of the data base connection (driver, host, port, dbname, user, password)}

\item{vector}{A boolean indicating whether the query should return a a vector that is the first column of a table}
}
\value{
A data frame containing the results of the redshift query (if vector=FALSE) or a vector containing the first column of the results (if vector=TRUE)
}
\description{
Open a connection to a SQL database, run the query, and close the connection.
}
\examples{
config <- list(driver="mydriver.redshift.amazonaws.com", port="5431", dbname="mydbname", user="kristorffeson", password="mYs3curePa33word!")
query_redshift('SELECT * FROM mytable', config)
}

