% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/minutiaer.R
\name{s}
\alias{s}
\title{Creates a String with interpolated values}
\usage{
s(text, envir = parent.frame())
}
\arguments{
\item{text}{A string}
}
\value{
The text string with the substituted interpolated variables
}
\description{
Creates a String with interpolated values
}
\examples{
name <- "Kristorffeson"
greeting <- s("Hi there, ${name}")
# => "Hi there, Kristorffeson"
}

